+++ library/Episciences/User/Tmp.php
@@ -27,12 +27,14 @@ class Episciences_User_Tmp extends Episciences_User
     }
 
     /**
-     * @param bool $forceInsert
+     * @param $forceInsert
      * @param bool $isCasRecording
+     * @param int $rvId
      * @return bool
      * @throws Zend_Db_Adapter_Exception
      */
-    public function save(bool $forceInsert = false, bool $isCasRecording = false): bool
+
+    public function save($forceInsert = false, bool $isCasRecording = false, int $rvId = RVID): bool
     {
         $db = Zend_Db_Table_Abstract::getDefaultAdapter();